# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Receitas API",
        "description": "API para gerenciamento de receitas culinárias com autenticação e CRUD completo",
        "contact": {
            "email": "admin@receitas.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor da API"
        }
    ],
    "paths": {
...
        "/api/receitas": {
            "get": {
                "tags": [
                    "Receitas"
                ],
                "summary": "Listar receitas do usuário",
                "description": "Retorna todas as receitas cadastradas pelo usuário autenticado",
                "operationId": "2c86ab6c0614cabe23053a1a724bdfd4",
                "responses": {
                    "200": {
                        "description": "Lista de receitas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nome": {
                                                "type": "string",
                                                "example": "Bolo de Chocolate"
                                            },
                                            "tempo_preparo_minutos": {
                                                "type": "integer",
                                                "example": 45
                                            },
                                            "porcoes": {
                                                "type": "integer",
                                                "example": 8
                                            },
                                            "ingredientes": {
                                                "type": "string",
                                                "example": "2 xícaras de farinha..."
                                            },
                                            "modo_preparo": {
                                                "type": "string",
                                                "example": "Misture os ingredientes..."
                                            },
                                            "id_categorias": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "id_usuarios": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Receitas"
                ],
                "summary": "Criar nova receita",
                "description": "Cria uma nova receita para o usuário autenticado",
                "operationId": "b39df697911a0211f65be7949aa217fb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "id_categorias",
                                    "tempo_preparo_minutos",
                                    "porcoes",
                                    "ingredientes",
                                    "modo_preparo"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Bolo de Chocolate"
                                    },
                                    "id_categorias": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tempo_preparo_minutos": {
                                        "type": "integer",
                                        "example": 45
                                    },
                                    "porcoes": {
                                        "type": "integer",
                                        "example": 8
                                    },
                                    "ingredientes": {
                                        "type": "string",
                                        "example": "2 xícaras de farinha..."
                                    },
                                    "modo_preparo": {
                                        "type": "string",
                                        "example": "Misture os ingredientes..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Receita criada com sucesso"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/receitas/{id}": {
            "get": {
                "tags": [
                    "Receitas"
                ],
                "summary": "Obter detalhes de uma receita",
                "description": "Retorna os detalhes de uma receita específica do usuário",
                "operationId": "fe3d6144cdf43f698397f09aeb16833b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da receita",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da receita"
                    },
                    "404": {
                        "description": "Receita não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Receitas"
                ],
                "summary": "Atualizar receita",
                "description": "Atualiza os dados de uma receita existente",
                "operationId": "ce24264a73b3d2c4269dbf441fc99729",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da receita",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Bolo de Chocolate"
                                    },
                                    "id_categorias": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tempo_preparo_minutos": {
                                        "type": "integer",
                                        "example": 45
                                    },
                                    "porcoes": {
                                        "type": "integer",
                                        "example": 8
                                    },
                                    "ingredientes": {
                                        "type": "string"
                                    },
                                    "modo_preparo": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Receita atualizada com sucesso"
                    },
                    "404": {
                        "description": "Receita não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Receitas"
                ],
                "summary": "Excluir receita",
                "description": "Remove uma receita do sistema",
                "operationId": "df52957822fbc930aa6f68c9e6dfefdb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da receita",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Receita excluída com sucesso"
                    },
                    "404": {
                        "description": "Receita não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/receitas/search/{term}": {
            "get": {
                "tags": [
                    "Receitas"
                ],
                "summary": "Pesquisar receitas",
                "description": "Busca receitas por nome, ingredientes ou modo de preparo",
                "operationId": "c311f8da5290ea35b3cf0eea0215cb00",
                "parameters": [
                    {
                        "name": "term",
                        "in": "path",
                        "description": "Termo de busca",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados da pesquisa"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/receitas/{id}/print": {
            "get": {
                "tags": [
                    "Receitas"
                ],
                "summary": "Gerar versão para impressão",
                "description": "Retorna HTML formatado para impressão da receita",
                "operationId": "37957bd39a80c797a1a94ca5e65e9062",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da receita",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "HTML para impressão"
                    },
                    "404": {
                        "description": "Receita não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}